Fighting Memory Leaks
Log4Net was a big affender so that's gone now.
Added some event unsubscribing, but that doesn't appear to be a big
problem.
The Lync SDK appears to have a leak of it's own, so there's no fixing
that, but it only occurs on login/logout of the client, so it's not a
huge problem.